pipeline/requirements.txt
scrapling
requests
beautifulsoup4
pillow
opencv-python-headless
supabase
python-dotenv
